requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.30.0
torch==2.0.1
pillow==9.5.0
python-multipart==0.0.6
numpy<2.0.0